--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "GIYA"]
-	path = GIYA
-	url = https://github.com/iceice666/GIYA.git
