+++ run.py
@@ -1,13 +1,20 @@
-
 import asyncio
+import importlib
+import sys
+from time import sleep
 
 
 async def main():
-    from src.main import Interface
+    Interface = importlib.import_module('.main', 'src.CLI').Interface
     await Interface().entrypoint()
+    return
 
 
-try:
-    asyncio.run(main())
-except:
-    pass
+asyncio.run(main())
+
+#loop = asyncio.new_event_loop()
+#loop.run_until_complete(main())
+
+sleep(3)
+
+sys.exit(0)
