+++ src/CLI/music.py
@@ -1,0 +1,103 @@
+import asyncio
+from typing import Optional
+
+from prompt_toolkit.application import in_terminal
+from rich.style import Style
+from rich.traceback import install
+
+from src.CLI.core import *
+from src.core.vlc_core import VLC
+
+install(show_locals=True)
+
+
+class Player:
+    flag_repeat: bool = False
+    flag_loop: bool = False
+    flag_skip: bool = False
+    _finish_signal_passed: bool = False
+
+    queue: list[Track] = []
+    nowplaying: Optional[Track] = None
+
+    def __init__(self):
+        self._rl = asyncio.get_running_loop()
+
+        _v = VLC()
+        _v._playing_end = self._playing_end
+
+        self.player = _v
+
+    def execute(self, cmd_args):
+        for c in cmd_args:
+            exec(c.replace("$", "self."))
+
+    def volume(self, vol: Optional[int] = None):
+        return self.player.volume(vol)
+
+    def position(self, pos: int | float):
+        self.player.position(pos)
+
+    def repeat(self):
+        self.flag_repeat = not self.flag_repeat
+
+    def loop(self):
+        self.flag_loop = not self.flag_loop
+
+    def skip(self):
+        self.player.stop()
+
+    def pause(self):
+        self.player.pause()
+
+    def resume(self):
+        self.player.resume()
+
+    def clear(self):
+        self.queue = []
+
+    async def add_track(self, track):
+
+        fetched_info = await NetworkIO.fetch_info(track)
+
+        self.queue += fetched_info
+
+        for i in fetched_info:
+            console.print(f'[Player] Added Track: {i.webpage_url}')
+
+        if not self.is_playing and self.nowplaying is None and len(self.queue) > 1:
+            await self.play()
+
+    async def play(self):
+        if not self.queue:
+            return
+        self.nowplaying = self.queue.pop(0)
+        if time.time() > self.nowplaying.expired_time:
+            console.print(
+                '[Player] [yellow]This link expired, re-fetching...[/yellow]')
+            self.nowplaying.source_url = (await NetworkIO.fetch_info(self.nowplaying))[0].source_url
+
+            self.nowplaying.expired_time = time.time() + 3600
+
+        self.player.set_uri(self.nowplaying.source_url)
+
+        async with in_terminal():
+            console.print('[Player] ', end='')
+            console.print(
+                'Nowplaying: ', self.nowplaying.title, style=Style(color='#D670B3'))
+
+        self.player.play()
+
+    async def _playing_end(self):
+        if self.flag_repeat and not self.flag_skip:
+            self.queue.insert(0, self.nowplaying)
+        elif self.flag_loop and not self.flag_skip:
+            self.queue.append(self.nowplaying)
+
+        self.nowplaying = None
+
+        await self.play()
+
+    @property
+    def is_playing(self):
+        return self.player.is_playing()
