--- src/cli/__init__.py
@@ -1,1 +1,0 @@
-from .Music import Commands
