+++ src/utils/utility.py
@@ -1,6 +1,5 @@
 import time
 
-
 def timeit(func):
     def wrapper(*args, **kwargs):
         start = time.time()
